+++ frontend/src/App.vue
@@ -143,9 +143,19 @@ export default {
     },
 
     created: function() {
+        // set backend URL
+        this.$root.backendHost = process.env.VUE_APP_BACKEND_HOST;
+        if (!this.$root.backendHost) {
+            const winLocation = window.location;
+            let port = '';
+            if (winLocation.port !== '' && `${winLocation.port}` !== '80' && `${winLocation.port}` !== '443') {
+                port = `:${winLocation.port}`;
+            }
+            this.$root.backendHost = `${winLocation.protocol}//${winLocation.hostname}${port}`;
+        }
+
         // configure neucore-js-client
-        ApiClient.instance.basePath =
-            `${window.location.protocol}//${window.location.hostname}:${window.location.port}/api`;
+        ApiClient.instance.basePath = `${this.$root.backendHost}/api`;
         ApiClient.instance.plugins = [superAgentPlugin(this, setCsrfHeader)];
 
         // initial route