+++ frontend/src/components/NavBar.vue
@@ -178,11 +178,11 @@ export default {
                 return;
             }
             this.selectedTheme = name;
-            const $enable = $(`head link[href*='dist/theme-${this.selectedTheme.toLowerCase()}']`);
+            const $enable = $(`head link[href*='css/theme-${this.selectedTheme.toLowerCase()}']`);
             if ($enable.attr('rel') === 'stylesheet') {
                 return;
             }
-            const $disable = $("head link[href*='dist/theme-']");
+            const $disable = $("head link[href*='css/theme-']");
             $disable.attr('rel', 'alternate stylesheet');
             $disable.attr('disabled', true);
             $enable.attr('rel', 'stylesheet');