+++ frontend/src/pages/SystemSettings--Directors.vue
@@ -65,11 +65,7 @@ export default {
         readSettings(this);
 
         // login URL for director chars
-        let port = '';
-        if (location.port !== "" && `${location.port}` !== '80' && `${location.port}` !== '443') {
-            port = `:${location.port}`;
-        }
-        this.loginUrlDirector = `${location.protocol}//${location.hostname}${port}/login-director`
+        this.loginUrlDirector = `${this.$root.backendHost}/login-director`
     },
 
     watch: {