+++ frontend/src/pages/SystemSettings--Mails.vue
@@ -3,7 +3,7 @@
     <div class="card-header">Sender</div>
     <div class="card-body">
         <span v-if="settings.mail_character === ''">
-            <a href="/login-mail"><img src="/static/eve_sso.png" alt="LOG IN with EVE Online"></a>
+            <a href="/login-mail"><img src="../assets/eve_sso.png" alt="LOG IN with EVE Online"></a>
         </span>
         <span v-else>
             <span class="text-info">{{ settings.mail_character }}</span>
@@ -249,7 +249,7 @@ function readSettings(vm) {
 }
 </script>
 
-<style type="text/scss" scoped>
+<style scoped>
     .input-resend {
         display: inline;
         width: 75px;