+++ frontend/src/superagent-plugin.js
@@ -10,6 +10,8 @@ function superAgentPlugin(vm, setCsrfHeader) {
     return function (request) {
         setCsrfHeader(vm, request);
 
+        request.withCredentials();
+
         vm.ajaxLoading(true);
         request.on('end', function () {
             vm.ajaxLoading(false);